Moved the code for the App Bar into its own folder.
Code for the SearchAppBar is now in its own folder, called 'AppBar.js'. It is also still being called from 'App.js'.
